[main] custom swal e ajustes scss

+++ src/app/layouts/content/content/content.component.ts
@@ -2,6 +2,8 @@ import { Component, HostListener, OnInit } from '@angular/core';
 import { LocalService } from '../../../shared/services/local.service';
 import { Router } from '@angular/router';
 import { ToastService } from '../../../shared/services/toast-service.service';
+import Swal from 'sweetalert2';
+import { SwalService } from '../../../shared/services/swal.service';
 
 @Component({
   selector: 'app-content',
@@ -17,7 +19,8 @@ export class ContentComponent implements OnInit {
   constructor(
     public localService: LocalService,
     private router: Router,
-    public toast: ToastService
+    public toast: ToastService,
+    public swal: SwalService
   ) {}
 
   ngOnInit() {
@@ -36,38 +39,30 @@ export class ContentComponent implements OnInit {
   changeLinkByScreenResize() {
     let links: any[];
 
-    if (this.isLargeScreen) {
-      links = [
-        {
-          name: 'Procurar',
-          icon: 'search',
-          routerLink: '/lifts',
-        },
-        {
-          name: this.localService.userIsDriver
-            ? 'Oferecer carona'
-            : 'Solicitar carona',
-          icon: 'add_circle',
-          routerLink: 'offer-lift',
-        },
-        {
-          name: 'Histórico',
-          icon: 'manage_search',
-          routerLink: '/lifts/history',
-        },
-      ];
-    } else {
-      links = [
-        {
-          icon: 'search',
-          routerLink: '/lifts',
-        },
-        {
-          icon: 'add_circle',
-          routerLink: 'offer-lift',
-        },
-      ];
-    }
+    links = [
+      {
+        name: 'Procurar',
+        icon: 'search',
+        routerLink: '/lifts',
+      },
+      {
+        name: this.localService.userIsDriver
+          ? 'Oferecer carona'
+          : 'Solicitar carona',
+        icon: 'add_circle',
+        routerLink: 'offer-lift',
+      },
+      {
+        name: 'Histórico',
+        icon: 'manage_search',
+        routerLink: '/lifts/history',
+      },
+      {
+        name: 'Carona atual',
+        icon: 'map',
+        routerLink: '/lifts/actual',
+      },
+    ];
 
     this.links = links;
   }
@@ -77,8 +72,20 @@ export class ContentComponent implements OnInit {
   }
 
   logout() {
-    this.localService.remove('user');
-    this.toast.showToastSucess('Deslogado com sucesso');
-    this.router.navigate(['/auth/login']);
+    Swal.fire({
+      icon: 'warning',
+      title: 'Deseja sair da sua conta?',
+      showCancelButton: true,
+      confirmButtonText: 'Sim',
+      cancelButtonText: `Não`,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.swal.showMessage('Deslogado com sucesso!', 'success');
+        this.localService.remove('user');
+        this.router.navigate(['/auth/login']);
+      } else if (result.isDismissed) {
+        this.swal.showMessage('Operação cancelada', 'error');
+      }
+    });
   }
 }
